requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
faiss-cpu
torch
pydantic
python-dotenv
python-multipart